AST-19525 - Packaging enforcement using the JWT
jwt-helper refactor
