+++ internal/wrappers/jwt-helper.go
@@ -1,5 +1,12 @@
 package wrappers
 
+import (
+	"encoding/json"
+	"strings"
+
+	"github.com/pkg/errors"
+)
+
 // JWTStruct model used to get all jwt fields
 type JWTStruct struct {
 	Exp            int      `json:"exp"`
@@ -71,4 +78,42 @@ type JWTWrapper interface {
 
 func NewJwtWrapper() JWTWrapper {
 	return &JWTStruct{}
+}
+
+// GetAllowedEngines will return a map with user allowed engines
+func (*JWTStruct) GetAllowedEngines() (allowedEngines map[string]bool, err error) {
+	accessToken, err := GetAccessToken()
+	if err != nil {
+		return nil, err
+	}
+	extractedToken, err := ExtractFromTokenToInterface(accessToken)
+	if err != nil {
+		return nil, errors.Errorf("Error extracting jwt - %v", err)
+	}
+	err = jwtToStruct(extractedToken, &JwtStruct)
+	if err != nil {
+		return nil, err
+	}
+	allowedEngines = fillBooleanMap(JwtStruct.AstLicense.LicenseData.AllowedEngines)
+	return allowedEngines, nil
+}
+
+func jwtToStruct(extractedToken interface{}, emptyJWT *JWTStruct) error {
+	marshaled, err := json.Marshal(extractedToken)
+	if err != nil {
+		return errors.Errorf("Error encoding jwt struct - %v", err)
+	}
+	err = json.Unmarshal(marshaled, &emptyJWT)
+	if err != nil {
+		return errors.Errorf("Error decoding jwt struct - %v", err)
+	}
+	return nil
+}
+
+func fillBooleanMap(engines []string) map[string]bool {
+	m := make(map[string]bool)
+	for _, value := range engines {
+		m[strings.ToLower(value)] = true
+	}
+	return m
 }
